+++ src/themes/global-theme-wrapper/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/markoarthofer22/react-components/compare/@markoarthofer22/react-components.global-theme-provider@1.2.0...@markoarthofer22/react-components.global-theme-provider@1.3.0) (2022-02-12)
+
+
+### Features
+
+* added alert components, added new vars to theme ([886a194](https://github.com/markoarthofer22/react-components/commit/886a19422948fde3c30335c12dd041b335e39935))
+
+
+
+
+
 # [1.2.0](https://github.com/markoarthofer22/react-components/compare/@markoarthofer22/react-components.global-theme-provider@1.1.2...@markoarthofer22/react-components.global-theme-provider@1.2.0) (2022-02-04)